Move discovery-related code into discovery.go.
